+++ src/kio/schema/add_partitions_to_txn/request/v2.py
@@ -8,6 +8,8 @@ from typing import ClassVar
 from kio.schema.entity import ProducerId
 from kio.schema.entity import TopicName
 from kio.schema.entity import TransactionalId
+from kio.schema.primitive import i16
+from kio.schema.primitive import i32
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
@@ -15,7 +17,7 @@ class AddPartitionsToTxnTopic:
     __flexible__: ClassVar[bool] = False
     name: TopicName = field(metadata={"kafka_type": "string"})
     """The name of the topic."""
-    partitions: tuple[int, ...] = field(metadata={"kafka_type": "int32"}, default=())
+    partitions: tuple[i32, ...] = field(metadata={"kafka_type": "int32"}, default=())
     """The partition indexes to add to the transaction"""
 
 
@@ -26,7 +28,7 @@ class AddPartitionsToTxnRequest:
     """The transactional id corresponding to the transaction."""
     producer_id: ProducerId = field(metadata={"kafka_type": "int64"})
     """Current producer id in use by the transactional id."""
-    producer_epoch: int = field(metadata={"kafka_type": "int16"})
+    producer_epoch: i16 = field(metadata={"kafka_type": "int16"})
     """Current epoch associated with the producer id."""
     topics: tuple[AddPartitionsToTxnTopic, ...]
     """The partitions to add to the transaction."""
