+++ src/kio/schema/alter_partition_reassignments/request/v0.py
@@ -7,12 +7,13 @@ from typing import ClassVar
 
 from kio.schema.entity import BrokerId
 from kio.schema.entity import TopicName
+from kio.schema.primitive import i32
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class ReassignablePartition:
     __flexible__: ClassVar[bool] = True
-    partition_index: int = field(metadata={"kafka_type": "int32"})
+    partition_index: i32 = field(metadata={"kafka_type": "int32"})
     """The partition index."""
     replicas: tuple[BrokerId, ...] = field(metadata={"kafka_type": "int32"}, default=())
     """The replicas to place the partitions on, or null to cancel a pending reassignment for this partition."""
@@ -30,7 +31,7 @@ class ReassignableTopic:
 @dataclass(frozen=True, slots=True, kw_only=True)
 class AlterPartitionReassignmentsRequest:
     __flexible__: ClassVar[bool] = True
-    timeout_ms: int = field(metadata={"kafka_type": "int32"}, default=60000)
+    timeout_ms: i32 = field(metadata={"kafka_type": "int32"}, default=i32(60000))
     """The time in ms to wait for the request to complete."""
     topics: tuple[ReassignableTopic, ...]
     """The topics to reassign."""
