+++ src/kio/schema/alter_user_scram_credentials/response/v0.py
@@ -5,13 +5,16 @@ from dataclasses import dataclass
 from dataclasses import field
 from typing import ClassVar
 
+from kio.schema.primitive import i16
+from kio.schema.primitive import i32
+
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class AlterUserScramCredentialsResult:
     __flexible__: ClassVar[bool] = True
     user: str = field(metadata={"kafka_type": "string"})
     """The user name."""
-    error_code: int = field(metadata={"kafka_type": "int16"})
+    error_code: i16 = field(metadata={"kafka_type": "int16"})
     """The error code."""
     error_message: str | None = field(metadata={"kafka_type": "string"})
     """The error message, if any."""
@@ -20,7 +23,7 @@ class AlterUserScramCredentialsResult:
 @dataclass(frozen=True, slots=True, kw_only=True)
 class AlterUserScramCredentialsResponse:
     __flexible__: ClassVar[bool] = True
-    throttle_time_ms: int = field(metadata={"kafka_type": "int32"})
+    throttle_time_ms: i32 = field(metadata={"kafka_type": "int32"})
     """The duration in milliseconds for which the request was throttled due to a quota violation, or zero if the request did not violate any quota."""
     results: tuple[AlterUserScramCredentialsResult, ...]
     """The results for deletions and alterations, one per affected user."""
