+++ src/kio/schema/create_acls/response/v2.py
@@ -5,11 +5,14 @@ from dataclasses import dataclass
 from dataclasses import field
 from typing import ClassVar
 
+from kio.schema.primitive import i16
+from kio.schema.primitive import i32
+
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class AclCreationResult:
     __flexible__: ClassVar[bool] = True
-    error_code: int = field(metadata={"kafka_type": "int16"})
+    error_code: i16 = field(metadata={"kafka_type": "int16"})
     """The result error, or zero if there was no error."""
     error_message: str | None = field(metadata={"kafka_type": "string"})
     """The result message, or null if there was no error."""
@@ -18,7 +21,7 @@ class AclCreationResult:
 @dataclass(frozen=True, slots=True, kw_only=True)
 class CreateAclsResponse:
     __flexible__: ClassVar[bool] = True
-    throttle_time_ms: int = field(metadata={"kafka_type": "int32"})
+    throttle_time_ms: i32 = field(metadata={"kafka_type": "int32"})
     """The duration in milliseconds for which the request was throttled due to a quota violation, or zero if the request did not violate any quota."""
     results: tuple[AclCreationResult, ...]
     """The results for each ACL creation."""
