+++ src/kio/schema/create_topics/response/v3.py
@@ -6,6 +6,8 @@ from dataclasses import field
 from typing import ClassVar
 
 from kio.schema.entity import TopicName
+from kio.schema.primitive import i16
+from kio.schema.primitive import i32
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
@@ -13,7 +15,7 @@ class CreatableTopicResult:
     __flexible__: ClassVar[bool] = False
     name: TopicName = field(metadata={"kafka_type": "string"})
     """The topic name."""
-    error_code: int = field(metadata={"kafka_type": "int16"})
+    error_code: i16 = field(metadata={"kafka_type": "int16"})
     """The error code, or 0 if there was no error."""
     error_message: str | None = field(metadata={"kafka_type": "string"})
     """The error message, or null if there was no error."""
@@ -22,7 +24,7 @@ class CreatableTopicResult:
 @dataclass(frozen=True, slots=True, kw_only=True)
 class CreateTopicsResponse:
     __flexible__: ClassVar[bool] = False
-    throttle_time_ms: int = field(metadata={"kafka_type": "int32"})
+    throttle_time_ms: i32 = field(metadata={"kafka_type": "int32"})
     """The duration in milliseconds for which the request was throttled due to a quota violation, or zero if the request did not violate any quota."""
     topics: tuple[CreatableTopicResult, ...]
     """Results for each topic we tried to create."""
