+++ src/kio/schema/describe_cluster/response/v0.py
@@ -6,6 +6,8 @@ from dataclasses import field
 from typing import ClassVar
 
 from kio.schema.entity import BrokerId
+from kio.schema.primitive import i16
+from kio.schema.primitive import i32
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
@@ -15,7 +17,7 @@ class DescribeClusterBroker:
     """The broker ID."""
     host: str = field(metadata={"kafka_type": "string"})
     """The broker hostname."""
-    port: int = field(metadata={"kafka_type": "int32"})
+    port: i32 = field(metadata={"kafka_type": "int32"})
     """The broker port."""
     rack: str | None = field(metadata={"kafka_type": "string"}, default=None)
     """The rack of the broker, or null if it has not been assigned to a rack."""
@@ -24,21 +26,21 @@ class DescribeClusterBroker:
 @dataclass(frozen=True, slots=True, kw_only=True)
 class DescribeClusterResponse:
     __flexible__: ClassVar[bool] = True
-    throttle_time_ms: int = field(metadata={"kafka_type": "int32"})
+    throttle_time_ms: i32 = field(metadata={"kafka_type": "int32"})
     """The duration in milliseconds for which the request was throttled due to a quota violation, or zero if the request did not violate any quota."""
-    error_code: int = field(metadata={"kafka_type": "int16"})
+    error_code: i16 = field(metadata={"kafka_type": "int16"})
     """The top-level error code, or 0 if there was no error"""
     error_message: str | None = field(metadata={"kafka_type": "string"}, default=None)
     """The top-level error message, or null if there was no error."""
     cluster_id: str = field(metadata={"kafka_type": "string"})
     """The cluster ID that responding broker belongs to."""
     controller_id: BrokerId = field(
-        metadata={"kafka_type": "int32"}, default=BrokerId(-1)
+        metadata={"kafka_type": "int32"}, default=BrokerId(i32(-1))
     )
     """The ID of the controller broker."""
     brokers: tuple[DescribeClusterBroker, ...]
     """Each broker in the response."""
-    cluster_authorized_operations: int = field(
-        metadata={"kafka_type": "int32"}, default=-2147483648
+    cluster_authorized_operations: i32 = field(
+        metadata={"kafka_type": "int32"}, default=i32(-2147483648)
     )
     """32-bit bitfield to represent authorized operations for this cluster."""
