+++ src/kio/schema/elect_leaders/request/v0.py
@@ -6,6 +6,7 @@ from dataclasses import field
 from typing import ClassVar
 
 from kio.schema.entity import TopicName
+from kio.schema.primitive import i32
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
@@ -13,7 +14,7 @@ class TopicPartitions:
     __flexible__: ClassVar[bool] = False
     topic: TopicName = field(metadata={"kafka_type": "string"})
     """The name of a topic."""
-    partitions: tuple[int, ...] = field(metadata={"kafka_type": "int32"}, default=())
+    partitions: tuple[i32, ...] = field(metadata={"kafka_type": "int32"}, default=())
     """The partitions of this topic whose leader should be elected."""
 
 
@@ -22,5 +23,5 @@ class ElectLeadersRequest:
     __flexible__: ClassVar[bool] = False
     topic_partitions: tuple[TopicPartitions, ...]
     """The topic partitions to elect leaders."""
-    timeout_ms: int = field(metadata={"kafka_type": "int32"}, default=60000)
+    timeout_ms: i32 = field(metadata={"kafka_type": "int32"}, default=i32(60000))
     """The time in ms to wait for the election to complete."""
